requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
defusedxml==0.7.1
Deprecated==1.2.18
fastapi==0.115.11
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
httpie==3.2.4
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
locust==2.33.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
opentelemetry-api==1.31.0
opentelemetry-distro==0.52b0
opentelemetry-exporter-otlp==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
packaging==24.2
protobuf==5.29.3
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PySocks==1.7.1
pyzmq==26.3.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==76.0.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
